refactored and passed rspecs
