+++ person.rb
@@ -1,0 +1,14 @@
+class Person
+  attr_accessor :name
+  attr_reader :age, :phase
+
+  def initialize(options = {})
+    @age = options.fetch(:age, 0)
+    @name = options.fetch(:name, "")  
+  end
+
+  def offer_high_five
+    "High five!"
+  end
+
+end
